data_summarization_agent/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
google-cloud-logging==3.5.0
pytest==7.4.0
httpx==0.24.1